requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
ctransformers
sentence_transformers
faiss_cpu
chainlit
huggingface_hub
langchain_community
streamlit
pillow
transformers
requests
googletrans==4.0.0-rc1
gTTS
google-generativeai